Restore correct test expectations #295
