+++ tests/extractcode/test_extract.py
@@ -51,13 +51,9 @@ class TestExtract(FileBasedTesting):
     def test_extract_file_function(self):
         test_file = self.get_test_loc('extract/basic_non_nested.tar.gz', copy=True)
         base = fileutils.parent_directory(test_file)
-        if on_linux or py2:
-            expected = ['a/b/a.txt', 'a/b/b.txt', 'a/c/c.txt']
-        if not on_linux and not py2:
-            expected = []
+        expected = ['a/b/a.txt', 'a/b/b.txt', 'a/c/c.txt']
         cleaned_test_file = test_file.replace(base, '')
-        if on_linux or py2:
-            expected_event = [
+        expected_event = [
             extract.ExtractEvent(
                 source=cleaned_test_file,
                 target=extractcode.get_extraction_path(cleaned_test_file),
@@ -69,8 +65,7 @@ class TestExtract(FileBasedTesting):
                 done=True, warnings=[], errors=[]
             )
         ]
-        if not on_linux and not py2:
-            expected_event = []
+
         target = extractcode.get_extraction_path(test_file)
         result = list(extract.extract_file(test_file, target))
         result = [r._replace(
@@ -82,14 +77,11 @@ class TestExtract(FileBasedTesting):
 
     def test_extract_archive_non_nested(self):
         test_dir = self.get_test_loc('extract/basic_non_nested.tar.gz', copy=True)
-        if on_linux or py2:
-            expected = (
+        expected = (
             'a/b/a.txt',
             'a/b/b.txt',
             'a/c/c.txt',
         )
-        if not on_linux and not py2:
-            expected = ()
         result = extract.extract(test_dir, recurse=False)
         check_no_error(result)
         check_files(extractcode.get_extraction_path(test_dir), expected)
@@ -124,13 +116,11 @@ class TestExtract(FileBasedTesting):
         """
         result = list(extract.extract(test_file, recurse=False))
         check_no_error(result)
-        if on_linux or py2:
-            expected = (
+
+        expected = (
             'somefilename-0.txt',
             'somefilename-1.txt',
         )
-        if not on_linux and not py2:
-            expected = ()
         test_dir = extractcode.get_extraction_path(test_file)
         check_files(test_dir, expected)
 
@@ -138,19 +128,15 @@ class TestExtract(FileBasedTesting):
         test_dir = self.get_test_loc('extract/readonly', copy=True)
         result = list(extract.extract(test_dir, recurse=False))
         check_no_error(result)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'read_only.tar.gz',
             'read_only.tar.gz-extract/somefilename-0.txt',
             'read_only.tar.gz-extract/somefilename-1.txt',
         ]
-        if not on_linux and not py2:
-            expected = ['read_only.tar.gz']
         check_files(test_dir, expected)
 
     def test_extract_tree_shallow_only(self):
-        if on_linux or py2:
-            expected = (
+        expected = (
             'a/a.tar.gz',
             'a/a.txt',
             'a/a.tar.gz-extract/a/b/a.txt',
@@ -191,15 +177,6 @@ class TestExtract(FileBasedTesting):
             'c/a.tar.gz-extract/a/b/b.txt',
             'c/a.tar.gz-extract/a/c/c.txt',
         )
-        if not on_linux and not py2:
-            expected = [
-            'a/a.tar.gz',
-            'a/a.txt',
-            'b/a.txt',
-            'b/b.tar.gz',
-            'c/a.tar.gz',
-            'c/a.txt'
-            ]
         test_dir = self.get_test_loc('extract/tree', copy=True)
         result = list(extract.extract(test_dir, recurse=False))
         check_no_error(result)
@@ -210,21 +187,18 @@ class TestExtract(FileBasedTesting):
         check_files(test_dir, expected)
 
     def test_extract_tree_recursive(self):
-        if on_linux or py2:
-            expected = (
+        expected = (
             'a/a.tar.gz',
             'a/a.txt',
             'a/a.tar.gz-extract/a/b/a.txt',
             'a/a.tar.gz-extract/a/b/b.txt',
             'a/a.tar.gz-extract/a/c/c.txt',
-
             'b/a.txt',
             'b/b.tar.gz',
             'b/b.tar.gz-extract/b/.svn/all-wcprops',
             'b/b.tar.gz-extract/b/.svn/entries',
             'b/b.tar.gz-extract/b/.svn/format',
             'b/b.tar.gz-extract/b/a/a.tar.gz',
-
             'b/b.tar.gz-extract/b/a/a.txt',
             'b/b.tar.gz-extract/b/a/.svn/all-wcprops',
             'b/b.tar.gz-extract/b/a/.svn/entries',
@@ -257,15 +231,6 @@ class TestExtract(FileBasedTesting):
             'c/a.tar.gz-extract/a/b/b.txt',
             'c/a.tar.gz-extract/a/c/c.txt',
         )
-        if not on_linux and not py2:
-            expected = [
-            'a/a.tar.gz',
-            'a/a.txt',
-            'b/a.txt',
-            'b/b.tar.gz',
-            'c/a.tar.gz',
-            'c/a.txt'
-            ]
         test_dir = self.get_test_loc('extract/tree', copy=True)
         result = list(extract.extract(test_dir, recurse=True))
         check_no_error(result)
@@ -276,8 +241,7 @@ class TestExtract(FileBasedTesting):
         check_files(test_dir, expected)
 
     def test_extract_tree_shallow_then_recursive(self):
-        if on_linux or py2:
-            shallow = (
+        shallow = (
             'a/a.tar.gz',
             'a/a.txt',
             'a/a.tar.gz-extract/a/b/a.txt',
@@ -290,7 +254,6 @@ class TestExtract(FileBasedTesting):
             'b/b.tar.gz-extract/b/.svn/entries',
             'b/b.tar.gz-extract/b/.svn/format',
             'b/b.tar.gz-extract/b/a/a.tar.gz',
-
             'b/b.tar.gz-extract/b/a/a.txt',
             'b/b.tar.gz-extract/b/a/.svn/all-wcprops',
             'b/b.tar.gz-extract/b/a/.svn/entries',
@@ -318,30 +281,18 @@ class TestExtract(FileBasedTesting):
             'c/a.tar.gz-extract/a/b/b.txt',
             'c/a.tar.gz-extract/a/c/c.txt',
         )
-        if not on_linux and not py2:
-            shallow = [
-            'a/a.tar.gz',
-            'a/a.txt',
-            'b/a.txt',
-            'b/b.tar.gz',
-            'c/a.tar.gz',
-            'c/a.txt'
-            ]
-        if on_linux or py2:
-            recursed = (
+        recursed = (
             'a/a.tar.gz',
             'a/a.txt',
             'a/a.tar.gz-extract/a/b/a.txt',
             'a/a.tar.gz-extract/a/b/b.txt',
             'a/a.tar.gz-extract/a/c/c.txt',
-
             'b/a.txt',
             'b/b.tar.gz',
             'b/b.tar.gz-extract/b/.svn/all-wcprops',
             'b/b.tar.gz-extract/b/.svn/entries',
             'b/b.tar.gz-extract/b/.svn/format',
             'b/b.tar.gz-extract/b/a/a.tar.gz',
-
             'b/b.tar.gz-extract/b/a/a.txt',
             'b/b.tar.gz-extract/b/a/.svn/all-wcprops',
             'b/b.tar.gz-extract/b/a/.svn/entries',
@@ -368,22 +319,12 @@ class TestExtract(FileBasedTesting):
             'b/b.tar.gz-extract/b/c/a.tar.gz-extract/a/b/a.txt',
             'b/b.tar.gz-extract/b/c/a.tar.gz-extract/a/b/b.txt',
             'b/b.tar.gz-extract/b/c/a.tar.gz-extract/a/c/c.txt',
-
             'c/a.tar.gz',
             'c/a.txt',
             'c/a.tar.gz-extract/a/b/a.txt',
             'c/a.tar.gz-extract/a/b/b.txt',
             'c/a.tar.gz-extract/a/c/c.txt',
         )
-        if not on_linux and not py2:
-            recursed = [
-            'a/a.tar.gz',
-            'a/a.txt',
-            'b/a.txt',
-            'b/b.tar.gz',
-            'c/a.tar.gz',
-            'c/a.txt'
-            ]
 
         test_dir = self.get_test_loc('extract/tree', copy=True)
         result = list(extract.extract(test_dir, recurse=False))
@@ -420,7 +361,7 @@ class TestExtract(FileBasedTesting):
                 assert isinstance(emsg, compat.unicode)
 
     @skipIf(py3 and not on_linux, 'Expectations are different on Windows and macOS')
-    def test_extract_tree_with_corrupted_archives(self):
+    def test_extract_tree_with_corrupted_archives_linux(self):
         expected = (
             'a.tar.gz',
         )
@@ -441,22 +382,16 @@ class TestExtract(FileBasedTesting):
         test_dir = self.get_test_loc('extract/corrupted', copy=True)
         result = list(extract.extract(test_dir, recurse=False))
         check_files(test_dir, expected)
-        if py2:
-            assert len(result) == 2
-            result = result[1]
-            errs = ['gzip decompression failed']
-            assert errs == result.errors
-            assert not result.warnings
-        else:
-            assert [] == result
+        assert len(result) == 2
+        result = result[1]
+        errs = ['gzip decompression failed']
+        assert errs == result.errors
+        assert not result.warnings
 
     def test_extract_with_empty_dir_and_small_files_ignores_empty_dirs(self):
-        if on_linux or py2:
-            expected = (
+        expected = (
             'empty_small.zip',
             'empty_small.zip-extract/empty_dirs_and_small_files/small_files/small_file.txt',)
-        if not on_linux and not py2:
-            expected = ['empty_small.zip']
         test_dir = self.get_test_loc('extract/small', copy=True)
         result = list(extract.extract(test_dir, recurse=True))
         check_no_error(result)
@@ -465,25 +400,18 @@ class TestExtract(FileBasedTesting):
     def test_extract_tar_with_broken_links(self):
         test_dir = self.get_test_loc('extract/broken_link', copy=True)
         result = list(extract.extract(test_dir, recurse=True))
-        if on_linux or py2:
-            expected = (
+        expected = (
             'broken-link.tar.bz2',
             'broken-link.tar.bz2-extract/openssl/test/Makefile',
         )
-        if not on_linux and not py2:
-            expected = ['broken-link.tar.bz2']
         check_files(test_dir, expected)
-        if on_linux or py2:
-            expected_warning = [[], []]
-        if not on_linux and not py2:
-            expected_warning = []
+        expected_warning = [[], []]
         warns = [r.warnings for r in result]
         assert expected_warning == warns
 
     def test_extract_nested_tar_file_recurse_only(self):
         test_file = self.get_test_loc('extract/nested/nested_tars.tar.gz', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'nested_tars.tar.gz',
             'nested_tars.tar.gz-extract/b/.svn/all-wcprops',
             'nested_tars.tar.gz-extract/b/.svn/entries',
@@ -516,16 +444,13 @@ class TestExtract(FileBasedTesting):
             'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/c/c.txt',
             'nested_tars.tar.gz-extract/b/c/a.txt'
         ]
-        if not on_linux and not py2:
-            expected = ['nested_tars.tar.gz']
         result = list(extract.extract(test_file, recurse=True))
         check_no_error(result)
         check_files(test_file, expected)
 
     def test_extract_nested_tar_file_shallow_only(self):
         test_dir = self.get_test_loc('extract/nested/nested_tars.tar.gz', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'nested_tars.tar.gz',
             'nested_tars.tar.gz-extract/b/.svn/all-wcprops',
             'nested_tars.tar.gz-extract/b/.svn/entries',
@@ -552,16 +477,13 @@ class TestExtract(FileBasedTesting):
             'nested_tars.tar.gz-extract/b/c/a.tar.gz',
             'nested_tars.tar.gz-extract/b/c/a.txt'
         ]
-        if not on_linux and not py2:
-            expected = ['nested_tars.tar.gz']
         result1 = list(extract.extract(test_dir, recurse=False))
         check_no_error(result1)
         check_files(test_dir, expected)
 
     def test_extract_nested_tar_file_shallow_then_recurse(self):
         test_file = self.get_test_loc('extract/nested/nested_tars.tar.gz', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'nested_tars.tar.gz',
             'nested_tars.tar.gz-extract/b/.svn/all-wcprops',
             'nested_tars.tar.gz-extract/b/.svn/entries',
@@ -582,9 +504,6 @@ class TestExtract(FileBasedTesting):
             'nested_tars.tar.gz-extract/b/b/.svn/format',
             'nested_tars.tar.gz-extract/b/b/.svn/text-base/a.txt.svn-base',
             'nested_tars.tar.gz-extract/b/b/a.txt',
-            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/a.txt',
-            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/b.txt',
-            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/c/c.txt',
             'nested_tars.tar.gz-extract/b/c/.svn/all-wcprops',
             'nested_tars.tar.gz-extract/b/c/.svn/entries',
             'nested_tars.tar.gz-extract/b/c/.svn/format',
@@ -592,10 +511,11 @@ class TestExtract(FileBasedTesting):
             'nested_tars.tar.gz-extract/b/c/.svn/text-base/a.tar.gz.svn-base',
             'nested_tars.tar.gz-extract/b/c/.svn/text-base/a.txt.svn-base',
             'nested_tars.tar.gz-extract/b/c/a.tar.gz',
+            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/a.txt',
+            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/b.txt',
+            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/c/c.txt',
             'nested_tars.tar.gz-extract/b/c/a.txt'
         ]
-        if not on_linux and not py2:
-            expected = ['nested_tars.tar.gz']
         result1 = list(extract.extract(test_file, recurse=False))
         result2 = list(extract.extract(test_file, recurse=True))
         check_no_error(result1)
@@ -604,8 +524,7 @@ class TestExtract(FileBasedTesting):
 
     def test_extract_dir_with_nested_tar_file_shallow_then_recurse(self):
         test_dir = self.get_test_loc('extract/nested', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'nested_tars.tar.gz',
             'nested_tars.tar.gz-extract/b/.svn/all-wcprops',
             'nested_tars.tar.gz-extract/b/.svn/entries',
@@ -626,9 +545,6 @@ class TestExtract(FileBasedTesting):
             'nested_tars.tar.gz-extract/b/b/.svn/format',
             'nested_tars.tar.gz-extract/b/b/.svn/text-base/a.txt.svn-base',
             'nested_tars.tar.gz-extract/b/b/a.txt',
-            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/a.txt',
-            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/b.txt',
-            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/c/c.txt',
             'nested_tars.tar.gz-extract/b/c/.svn/all-wcprops',
             'nested_tars.tar.gz-extract/b/c/.svn/entries',
             'nested_tars.tar.gz-extract/b/c/.svn/format',
@@ -636,10 +552,11 @@ class TestExtract(FileBasedTesting):
             'nested_tars.tar.gz-extract/b/c/.svn/text-base/a.tar.gz.svn-base',
             'nested_tars.tar.gz-extract/b/c/.svn/text-base/a.txt.svn-base',
             'nested_tars.tar.gz-extract/b/c/a.tar.gz',
+            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/a.txt',
+            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/b/b.txt',
+            'nested_tars.tar.gz-extract/b/c/a.tar.gz-extract/a/c/c.txt',
             'nested_tars.tar.gz-extract/b/c/a.txt'
         ]
-        if not on_linux and not py2:
-            expected = ['nested_tars.tar.gz']
         result1 = list(extract.extract(test_dir, recurse=False))
         result2 = list(extract.extract(test_dir, recurse=True))
         check_no_error(result1)
@@ -648,36 +565,29 @@ class TestExtract(FileBasedTesting):
 
     def test_extract_zip_with_spaces_in_name(self):
         test_dir = self.get_test_loc('extract/space-zip', copy=True)
-        if on_linux or py2:
-            expected = (
+        expected = (
             'with spaces in name.zip',
             'with spaces in name.zip-extract/empty_dirs_and_small_files/small_files/small_file.txt'
         )
-        if not on_linux and not py2:
-            expected = ['with spaces in name.zip']
         result = list(extract.extract(test_dir, recurse=True))
         check_no_error(result)
         check_files(test_dir, expected)
 
     def test_extract_tar_gz_with_spaces_in_name(self):
         test_dir = self.get_test_loc('extract/space-tgz', copy=True)
-        if on_linux or py2:
-            expected = (
+        expected = (
             'with spaces in name.tar.gz',
             'with spaces in name.tar.gz-extract/a/b/a.txt',
             'with spaces in name.tar.gz-extract/a/b/b.txt',
             'with spaces in name.tar.gz-extract/a/c/c.txt',
         )
-        if not on_linux and not py2:
-            expected = ['with spaces in name.tar.gz']
         result = list(extract.extract(test_dir, recurse=True))
         check_no_error(result)
         check_files(test_dir, expected)
 
     def test_extract_tar_with_special_files(self):
         test_dir = self.get_test_loc('extract/special', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             't.tgz',
             't.tgz-extract/0-REGTYPE',
             't.tgz-extract/0-REGTYPE-TEXT',
@@ -685,8 +595,6 @@ class TestExtract(FileBasedTesting):
             't.tgz-extract/S-SPARSE',
             't.tgz-extract/S-SPARSE-WITH-NULLS',
         ]
-        if not on_linux and not py2:
-            expected = ['t.tgz']
         result = list(extract.extract(test_dir, recurse=True))
         check_files(test_dir, expected)
 
@@ -961,14 +869,11 @@ class TestExtract(FileBasedTesting):
 
     def test_extract_nested_arch_with_corrupted_compressed_should_extract_inner_archives_only_once(self):
         test_file = self.get_test_loc('extract/nested_not_compressed/nested_with_not_compressed_gz_file.tgz', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'nested_with_not_compressed_gz_file.tgz',
             'nested_with_not_compressed_gz_file.tgz-extract/top/file',
             'nested_with_not_compressed_gz_file.tgz-extract/top/notcompressed.gz'
         ]
-        if not on_linux and not py2:
-            expected = ['nested_with_not_compressed_gz_file.tgz']
         result = list(extract.extract(test_file, recurse=True))
         check_no_error(result)
         check_files(test_file, expected)
@@ -976,8 +881,7 @@ class TestExtract(FileBasedTesting):
     def test_extract_directory_with_office_docs(self):
         test_dir = self.get_test_loc('extract/office_docs', copy=True)
         result = list(extract.extract(test_dir, kinds=(extractcode.docs,), recurse=True))
-        if on_linux or py2:
-            expected = [
+        expected = [
             'abc.docx',
             'abc.docx-extract/[Content_Types].xml',
             'abc.docx-extract/docProps/app.xml',
@@ -1060,14 +964,12 @@ class TestExtract(FileBasedTesting):
             'ppt.pptx-extract/ppt/tableStyles.xml',
             'ppt.pptx-extract/ppt/media/image1.png'
         ]
-        if not on_linux and not py2:
-            expected = ['abc.docx', 'excel.xlsx', 'ppt.pptx']
         check_files(test_dir, expected)
         check_no_error(result)
 
     def touch(self, location):
         with open(location, 'wb') as f:
-            f.write(b'\n')
+            f.write('\n')
 
     def fake_extract(self, location):
         extracted = os.path.join(location + 'extract')
@@ -1122,10 +1024,7 @@ class TestExtract(FileBasedTesting):
         test_src_file = join(test_src_dir, 'basic.zip')
         test_tgt_dir = join(scancode_root, test_src_file) + extractcode.EXTRACT_SUFFIX
         result = list(extract.extract(test_src_file))
-        if on_linux or py2:
-            expected = ['c/a/a.txt', 'c/b/a.txt', 'c/c/a.txt']
-        if not on_linux and not py2:
-            expected = []
+        expected = ['c/a/a.txt', 'c/b/a.txt', 'c/c/a.txt']
         check_files(test_tgt_dir, expected)
         for r in result:
             assert [] == r.warnings
@@ -1137,8 +1036,7 @@ class TestExtract(FileBasedTesting):
     @skipIf(on_windows, 'Windows behavior is slightly different with relative paths')
     def test_extract_zipslip_tar_posix(self):
         test_dir = self.get_test_loc('extract/zipslip', copy=True)
-        if on_linux or py2:
-            expected = [
+        expected = [
             'README.md',
             'origin.ABOUT',
             'zip-slip-win.tar',
@@ -1153,15 +1051,6 @@ class TestExtract(FileBasedTesting):
             'zip-slip.zip',
             'zip-slip.zip-extract/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/dotdot/tmp/evil.txt',
             'zip-slip.zip-extract/good.txt'
-        ]
-        if not on_linux and not py2:
-            expected = [
-            'README.md',
-            'origin.ABOUT',
-            'zip-slip-win.tar',
-            'zip-slip-win.zip',
-            'zip-slip.tar',
-            'zip-slip.zip'
         ]
 
         result = list(extract.extract(test_dir, recurse=True))
